+++ .ipython/profile_default/startup/00-imports.py
@@ -1,10 +1,9 @@
 """Preload common project imports for the local IPython shell."""
 
-from agent.create_ansible_playbook import create_ansible_playbook  # noqa: F401
-from agent.generate_playbook import GeneratePlaybookAgent  # noqa: F401
-from agent.orchestrator import OrchestratorAgent  # noqa: F401
-from agent.playbook_metadata import PlaybookMetadataAgent  # noqa: F401
-from agent.tools import (  # noqa: F401
+from devops_bot.agents.orchestrator import OrchestratorAgent  # noqa: F401
+from devops_bot.agents.playbook_generator import GeneratePlaybookAgent  # noqa: F401
+from devops_bot.agents.playbook_metadata import PlaybookMetadataAgent  # noqa: F401
+from devops_bot.tools import (  # noqa: F401
     create_git_branch,
     create_git_commit,
     get_ansible_inventory_groups,
@@ -16,6 +15,7 @@ from agent.tools import (  # noqa: F401
     run_ansible_playbook,
     search_web,
 )
+from devops_bot.tools.playbooks import create_ansible_playbook  # noqa: F401
 
 print("Loaded agents: OrchestratorAgent, GeneratePlaybookAgent, PlaybookMetadataAgent")
 print(
