+++ devops_bot/agents/__init__.py
@@ -1,0 +1,1 @@
+"""LLM-backed agents used by the DevOps bot."""
