+++ devops_bot/agents/playbook_generator.py
@@ -1,10 +1,10 @@
 import yaml
 from pydantic import BaseModel, Field, field_validator
 
-from .run_history import record_event
-from .tools import get_ansible_inventory_groups, http_get, search_web
-from .tools.ansible import check_ansible_playbook_syntax
-from .utils import build_agent, build_model
+from ..factory import build_agent, build_model
+from ..history import record_event
+from ..tools.ansible import check_ansible_playbook_syntax, get_ansible_inventory_groups
+from ..tools.web import http_get, search_web
 
 SYSTEM_PROMPT = """
 ## Execution Contract
