+++ devops_bot/tools/__init__.py
@@ -16,9 +16,9 @@ __all__ = [
     "create_git_branch",
     "create_git_commit",
     "get_ansible_inventory_groups",
+    "get_ansible_playbook_registry",
     "git_push",
     "git_status",
-    "get_ansible_playbook_registry",
     "http_get",
     "list_git_commits",
     "run_ansible_playbook",
