+++ devops_bot/tools/playbooks.py
@@ -1,13 +1,23 @@
 import re
 from pathlib import Path
+from typing import Callable, Protocol, TypedDict
 
 import yaml
 from pydantic import BaseModel
 from strands import tool
 
-from .generate_playbook import GeneratePlaybookAgent
-from .playbook_metadata import GeneratedPlaybookMetadata, PlaybookMetadataAgent
-from .run_history import record_event
+from ..agents.playbook_editor import (
+    EditAnsiblePlaybookAgent,
+    EditedAnsiblePlaybook,
+    validate_playbook_file_content,
+)
+from ..agents.playbook_generator import GeneratePlaybookAgent
+from ..agents.playbook_metadata import GeneratedPlaybookMetadata, PlaybookMetadataAgent
+from ..history import record_event
+from .ansible import (
+    check_ansible_playbook_syntax,
+    get_ansible_playbook_registry,
+)
 
 PLAYBOOKS_DIR = Path("ansible/playbooks")
 
@@ -30,13 +40,13 @@ def create_ansible_playbook(query: str) -> Path | None:
     Returns:
         The path to the written playbook under ansible/playbooks, or None if declined
     """
-    result = CreateAnsiblePlaybookWorkflow().run(query)
+    result = CreateAnsiblePlaybook().run(query)
     if not result.written:
         return None
     return result.path
 
 
-class CreateAnsiblePlaybookWorkflow:
+class CreateAnsiblePlaybook:
     def __init__(
         self,
         *,
@@ -104,8 +114,7 @@ class CreateAnsiblePlaybookWorkflow:
                 status="declined",
                 what="Playbook write was declined.",
                 why=(
-                    "The workflow requires explicit confirmation before "
-                    "creating a new playbook file."
+                    "The tool requires explicit confirmation before creating a new playbook file."
                 ),
                 details={"path": str(playbook_path), "approved": False},
             )
@@ -203,3 +212,177 @@ def summarize_generated_playbook(yaml_text: str) -> dict[str, str | int | list[s
                 task_count += len(tasks)
 
     return {"hosts": hosts, "task_count": task_count}
+
+
+class EditAnsiblePlaybookResult(BaseModel):
+    path: Path
+    summary: str
+    edited_content: str
+    written: bool
+    requires_remote_rerun: bool
+    syntax_check_passed: bool
+
+
+class EditAnsiblePlaybookToolResult(TypedDict):
+    path: str
+    summary: str
+    written: bool
+    requires_remote_rerun: bool
+    syntax_check_passed: bool
+
+
+class AnsiblePlaybookEditor(Protocol):
+    def run(
+        self,
+        *,
+        playbook_path: Path,
+        current_content: str,
+        requested_change: str,
+    ) -> EditedAnsiblePlaybook: ...
+
+
+@tool
+def edit_ansible_playbook(
+    playbook_path: str,
+    requested_change: str,
+) -> EditAnsiblePlaybookToolResult:
+    """
+    Edit a validated Ansible playbook locally and syntax-check the result.
+
+    Args:
+        playbook_path: Relative path to an existing playbook in the validated registry.
+        requested_change: Natural-language description of the local edit to make.
+
+    Returns:
+        A dictionary describing whether the syntax-checked edit was written.
+    """
+    result = EditAnsiblePlaybook().run(
+        playbook_path=playbook_path,
+        requested_change=requested_change,
+    )
+    return {
+        "path": str(result.path),
+        "summary": result.summary,
+        "written": result.written,
+        "requires_remote_rerun": result.requires_remote_rerun,
+        "syntax_check_passed": result.syntax_check_passed,
+    }
+
+
+class EditAnsiblePlaybook:
+    def __init__(
+        self,
+        *,
+        editor: AnsiblePlaybookEditor | None = None,
+        syntax_checker: Callable[[str], None] = check_ansible_playbook_syntax,
+    ) -> None:
+        self.editor = editor or EditAnsiblePlaybookAgent()
+        self.syntax_checker = syntax_checker
+
+    def run(self, *, playbook_path: str, requested_change: str) -> EditAnsiblePlaybookResult:
+        target_path = _validate_registry_playbook_path(playbook_path)
+        record_event(
+            kind="playbook_edit_started",
+            status="started",
+            what=f"Started local edit for playbook `{playbook_path}`.",
+            why="Repair an existing validated playbook before considering another remote run.",
+            details={"path": playbook_path, "requested_change": requested_change},
+        )
+
+        current_content = target_path.read_text(encoding="utf-8")
+        edited = self.editor.run(
+            playbook_path=target_path,
+            current_content=current_content,
+            requested_change=requested_change,
+        )
+        self.syntax_checker(edited.content)
+        validate_playbook_file_content(edited.content)
+
+        print_edit_preview(path=target_path, edited=edited)
+        record_event(
+            kind="playbook_edit_preview_presented",
+            status="completed",
+            what="Presented the edited playbook preview.",
+            why="Show the local edit summary before asking for write approval.",
+            details={
+                "path": str(target_path),
+                "summary": edited.summary,
+                "requires_remote_rerun": edited.requires_remote_rerun,
+            },
+        )
+
+        written = confirm_edit(target_path)
+        if not written:
+            print("Playbook edit not written.")
+            record_event(
+                kind="playbook_edit_declined",
+                status="declined",
+                what=f"Declined local edit for playbook `{playbook_path}`.",
+                why="The tool requires explicit confirmation before editing a playbook file.",
+                details={"path": str(target_path), "approved": False},
+            )
+            return EditAnsiblePlaybookResult(
+                path=target_path,
+                summary=edited.summary,
+                edited_content=edited.content,
+                written=False,
+                requires_remote_rerun=edited.requires_remote_rerun,
+                syntax_check_passed=True,
+            )
+
+        target_path.write_text(_normalize_file_content(edited.content), encoding="utf-8")
+        print(f"Edited playbook at {target_path}.")
+        record_event(
+            kind="playbook_edit_written",
+            status="completed",
+            what=f"Wrote local edit for playbook `{playbook_path}`.",
+            why="Persist the syntax-checked playbook repair under ansible/playbooks.",
+            details={
+                "path": str(target_path),
+                "approved": True,
+                "summary": edited.summary,
+                "requires_remote_rerun": edited.requires_remote_rerun,
+                "syntax_check_passed": True,
+            },
+        )
+        return EditAnsiblePlaybookResult(
+            path=target_path,
+            summary=edited.summary,
+            edited_content=edited.content,
+            written=True,
+            requires_remote_rerun=edited.requires_remote_rerun,
+            syntax_check_passed=True,
+        )
+
+
+def _validate_registry_playbook_path(playbook_path: str) -> Path:
+    requested_path = Path(playbook_path)
+    if requested_path.is_absolute():
+        raise ValueError("playbook path must be relative")
+
+    registry_paths = {entry["path"] for entry in get_ansible_playbook_registry()}
+    if playbook_path not in registry_paths:
+        raise ValueError(f"Playbook path is not in the registry: {playbook_path}")
+
+    resolved_playbooks_dir = PLAYBOOKS_DIR.resolve()
+    resolved_path = requested_path.resolve()
+    if resolved_playbooks_dir not in resolved_path.parents:
+        raise ValueError(f"Playbook path must be under {PLAYBOOKS_DIR}: {playbook_path}")
+    return requested_path
+
+
+def confirm_edit(path: Path) -> bool:
+    response = input(f"Write edited playbook to {path}? [y/N]: ").strip().lower()
+    return response in {"y", "yes"}
+
+
+def print_edit_preview(*, path: Path, edited: EditedAnsiblePlaybook) -> None:
+    print(f"Path: {path}")
+    print(f"Summary: {edited.summary}")
+    print(f"Requires remote rerun: {str(edited.requires_remote_rerun).lower()}")
+    print()
+    print(edited.content)
+
+
+def _normalize_file_content(content: str) -> str:
+    return f"{content.rstrip()}\n"
