+++ tests/tools/test_git.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from agent.tools import (
+from devops_bot.tools import (
     create_git_branch,
     create_git_commit,
     git_push,
